+++ modules/PSDev/commands/docusaurus.sidebar.js
@@ -10,8 +10,21 @@
 
 module.exports = [
     'PSDev/commands/Add-NumberFormater',
+    'PSDev/commands/Convert-Object',
     'PSDev/commands/Debug-String',
+    'PSDev/commands/Get-Color',
+    'PSDev/commands/Get-DotNetVersion',
     'PSDev/commands/Get-Office365IPURL',
+    'PSDev/commands/Get-PublicIP',
+    'PSDev/commands/Get-StringHash',
+    'PSDev/commands/Group-ObjectEvenly',
+    'PSDev/commands/New-Password',
+    'PSDev/commands/Resolve-IPinSubnet',
+    'PSDev/commands/Set-EnvironmentVariable',
+    'PSDev/commands/Switch-Object',
+    'PSDev/commands/Test-AllHashKeysAreTrue',
     'PSDev/commands/Test-Office365IPURL',
-    'PSDev/commands/Test-PSGalleryNameAvailability'
+    'PSDev/commands/Test-PasswordAgainstPwnedPasswordService',
+    'PSDev/commands/Test-PSGalleryNameAvailability',
+    'PSDev/commands/Test-RebootPending'
 ];