+++ modules/PSDev/changelog.md
@@ -9,6 +9,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added function Convert-Object
+- Added function Get-Color
+- Added function Get-DotNetVersion
+- Added function Get-PublicIP
+- Added function Get-StringHash
+- Added function Group-ObjectEvenly
+- Added function New-Password
+- Added function Resolve-IPinSubnet
+- Added function Set-EnvironmentVariable
+- Added function Switch-Object
+- Added function Test-AllHashKeysAreTrue
+- Added function Test-RebootPending
+- Added function Test-PasswordAgainstPwnedPasswordService
+
+### Fixed
+
+- Fix: Convert-Object: Changed securestring to string length evaulation. Length is different on windows and unix based systems so a fixed length value can not be evaluated.
+- Fix: Test-PasswordAgainstPwnedPasswordService: Splitting response array on carrage return and linefeed only works on windows. On unix only a linefeed is used and string is not splitted correctly.
+
 ## [1.0.3] - 2022-12-22
 
 ### Fixed